requirements.txt
genesis-world
--find-links https://download.pytorch.org/whl/torch_stable.html
torch
torchvision
redis
pandas
boto3
botocore
python-dotenv
sqlalchemy
psycopg2
random_word